Try harder to generate unique discriminators (#198)
Guarantee unique discriminators using rand::seq::index::sample

+++ rust/usernames/src/constants.rs
@@ -35,7 +35,7 @@ const COMPRESSED_BASE_POINTS_RAW: [[u8; 32]; 3] = [
 
 // 37^48 will overflow the Scalar. See nickname_scalar implementation for details.
 pub(crate) const MAX_NICKNAME_LENGTH: usize = 48;
-pub(crate) const DISCRIMINATOR_RANGES: [Range<u32>; 8] = [
+pub(crate) const DISCRIMINATOR_RANGES: [Range<usize>; 8] = [
     1..100,
     100..1_000,
     1_000..10_000,
@@ -46,7 +46,7 @@ pub(crate) const DISCRIMINATOR_RANGES: [Range<u32>; 8] = [
     100_000_000..1_000_000_000,
 ];
 
-pub(crate) const CANDIDATES_PER_RANGE: [u8; 8] = [4, 3, 3, 2, 2, 2, 2, 2];
+pub(crate) const CANDIDATES_PER_RANGE: [usize; 8] = [4, 3, 3, 2, 2, 2, 2, 2];
 
 #[cfg(test)]
 mod test {
